blobstor: drop compressor
1. It's very error-prone.
2. It's very ineffective for ranged requests.
3. It spends quite some CPU and memory.
