+++ cmd/neofs-adm/internal/modules/storagecfg/config.go
@@ -50,10 +50,6 @@ storage:
       blobstor:
         perm: 0644  # permissions for blobstor files(directories: +x for current user and group)
         depth: 2  # max depth of object tree storage in FS
-        compress: true  # turn on/off Zstandard compression (level 3) of stored objects
-        compression_exclude_content_types:
-          - audio/*
-          - video/*
 
       gc:
         remover_batch_size: 200  # number of objects to be removed by the garbage collector
