+++ config/example/node.env
@@ -103,9 +103,6 @@ NEOFS_STORAGE_SHARDS_0_METABASE_MAX_BATCH_SIZE=100
 NEOFS_STORAGE_SHARDS_0_METABASE_MAX_BATCH_DELAY=10ms
 NEOFS_STORAGE_SHARDS_0_METABASE_SEARCH_ITERATION_LIMIT=0
 ### Blobstor config
-NEOFS_STORAGE_SHARDS_0_COMPRESS=true
-NEOFS_STORAGE_SHARDS_0_COMPRESSION_EXCLUDE_CONTENT_TYPES="audio/* video/*"
-### FSTree config
 NEOFS_STORAGE_SHARDS_0_BLOBSTOR_TYPE=fstree
 NEOFS_STORAGE_SHARDS_0_BLOBSTOR_PATH=tmp/0/blob
 NEOFS_STORAGE_SHARDS_0_BLOBSTOR_PERM=0644
@@ -130,8 +127,6 @@ NEOFS_STORAGE_SHARDS_1_METABASE_MAX_BATCH_SIZE=200
 NEOFS_STORAGE_SHARDS_1_METABASE_MAX_BATCH_DELAY=20ms
 NEOFS_STORAGE_SHARDS_1_METABASE_SEARCH_ITERATION_LIMIT=5000
 ### Blobstor config
-NEOFS_STORAGE_SHARDS_1_COMPRESS=false
-### FSTree config
 NEOFS_STORAGE_SHARDS_1_BLOBSTOR_TYPE=fstree
 NEOFS_STORAGE_SHARDS_1_BLOBSTOR_PATH=tmp/1/blob
 NEOFS_STORAGE_SHARDS_1_BLOBSTOR_PERM=0644
