+++ config/example/node.yaml
@@ -123,8 +123,6 @@ storage:
       max_batch_delay: 20ms
       search_iteration_limit: 5000 # max number of consecutive objects to be iterated without a match during search operations; 0 means no limit, default to 10000
 
-    compress: false  # turn on/off zstd(level 3) compression of stored objects
-
     blobstor:
       perm: 0644  # permissions for blobstor files(directories: +x for current user and group)
       depth: 5  # max depth of object tree storage in FS
@@ -152,11 +150,6 @@ storage:
         max_batch_size: 100
         max_batch_delay: 10ms
         search_iteration_limit: 0 # max number of consecutive objects to be iterated without a match during search operations; 0 means no limit, default to 10000
-
-      compress: true  # turn on/off zstd(level 3) compression of stored objects
-      compression_exclude_content_types:
-        - audio/*
-        - video/*
 
       blobstor:
         type: fstree
