+++ docs/storage-node-configuration.md
@@ -158,8 +158,6 @@ The following table describes configuration for each shard.
 
 | Parameter                           | Type                                         | Default value | Description                                                                                                                                                                                                       |
 |-------------------------------------|----------------------------------------------|---------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
-| `compress`                          | `bool`                                       | `false`       | Flag to enable compression.                                                                                                                                                                                       |
-| `compression_exclude_content_types` | `[]string`                                   |               | List of content-types to disable compression for. Content-type is taken from `Content-Type` object attribute. Each element can contain a star `*` as a first (last) character, which matches any prefix (suffix). |
 | `mode`                              | `string`                                     | `read-write`  | Shard Mode.<br/>Possible values:  `read-write`, `read-only`, `degraded`, `degraded-read-only`, `disabled`                                                                                                         |
 | `writecache`                        | [Writecache config](#writecache-subsection)  |               | Write-cache configuration.                                                                                                                                                                                        |
 | `metabase`                          | [Metabase config](#metabase-subsection)      |               | Metabase configuration.                                                                                                                                                                                           |
