+++ pkg/local_object_storage/blobstor/fstree/head_bench_test.go
@@ -9,31 +9,17 @@ import (
 func BenchmarkFSTree_HeadVsGet(b *testing.B) {
 	for _, size := range payloadSizes {
 		b.Run(generateSizeLabel(size), func(b *testing.B) {
-			runHeadVsGetBenchmark(b, size, false)
+			runHeadVsGetBenchmark(b, size)
 		})
 	}
 }
 
-func BenchmarkFSTree_HeadVsGet_Compressed(b *testing.B) {
-	for _, size := range payloadSizes {
-		b.Run(generateSizeLabel(size), func(b *testing.B) {
-			runHeadVsGetBenchmark(b, size, true)
-		})
-	}
-}
-
-func runHeadVsGetBenchmark(b *testing.B, payloadSize int, compressed bool) {
+func runHeadVsGetBenchmark(b *testing.B, payloadSize int) {
 	fsTree := setupFSTree(b)
-
-	suffix := ""
-	if compressed {
-		setupCompressor(b, fsTree)
-		suffix = "_Compressed"
-	}
 
 	addr := prepareSingleObject(b, fsTree, payloadSize)
 
-	b.Run("Head"+suffix, func(b *testing.B) {
+	b.Run("Head", func(b *testing.B) {
 		b.ReportAllocs()
 		for b.Loop() {
 			_, err := fsTree.Head(addr)
@@ -43,7 +29,7 @@ func runHeadVsGetBenchmark(b *testing.B, payloadSize int, compressed bool) {
 		}
 	})
 
-	b.Run("ReadHeader"+suffix, func(b *testing.B) {
+	b.Run("ReadHeader", func(b *testing.B) {
 		buf := make([]byte, iobject.NonPayloadFieldsBufferLength*2)
 
 		b.ReportAllocs()
@@ -55,7 +41,7 @@ func runHeadVsGetBenchmark(b *testing.B, payloadSize int, compressed bool) {
 		}
 	})
 
-	b.Run("Get"+suffix, func(b *testing.B) {
+	b.Run("Get", func(b *testing.B) {
 		b.ReportAllocs()
 		for b.Loop() {
 			_, err := fsTree.Get(addr)
