+++ pkg/local_object_storage/blobstor/fstree/util.go
@@ -119,3 +119,18 @@ func (l *limitedFileReader) Seek(offset int64, whence int) (int64, error) {
 	}
 	return 0, err
 }
+
+func isCompressed(data []byte) bool {
+	// zstdFrameMagic contains first 4 bytes of any compressed object
+	// https://github.com/klauspost/compress/blob/master/zstd/framedec.go#L58 .
+	var zstdFrameMagic = []byte{0x28, 0xb5, 0x2f, 0xfd}
+
+	return len(data) >= 4 && bytes.Equal(data[:4], zstdFrameMagic)
+}
+
+func decompress(data []byte) ([]byte, error) {
+	if !isCompressed(data) {
+		return data, nil
+	}
+	return zstd.DecodeTo(nil, data)
+}
