+++ pkg/local_object_storage/shard/exists_test.go
@@ -7,8 +7,10 @@ import (
 
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard/mode"
+	cidtest "github.com/nspcc-dev/neofs-sdk-go/container/id/test"
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
+	usertest "github.com/nspcc-dev/neofs-sdk-go/user/test"
 	"github.com/stretchr/testify/require"
 )
 
@@ -64,4 +66,19 @@ func TestExists(t *testing.T) {
 		require.Error(t, err)
 		require.False(t, res)
 	})
+}
+
+func testObject(sz uint64) *object.Object {
+	raw := object.New(cidtest.ID(), usertest.ID())
+
+	raw.SetID(oidtest.ID())
+	raw.SetPayload(make([]byte, sz))
+
+	// fit the binary size to the required
+	data := raw.Marshal()
+	if ln := uint64(len(data)); ln > sz {
+		raw.SetPayload(raw.Payload()[:sz-(ln-sz)])
+	}
+
+	return raw
 }
