+++ pkg/local_object_storage/shard/generic_test.go
@@ -7,7 +7,6 @@ import (
 	"testing"
 
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/common"
-	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/compression"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/blobstor/fstree"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/internal/storagetest"
 	"github.com/nspcc-dev/neofs-node/pkg/local_object_storage/shard/mode"
@@ -52,10 +51,6 @@ func TestBlobstorGeneric(t *testing.T) {
 		fsTree := fstree.New(
 			fstree.WithPath(filepath.Join(dir, "fstree")),
 			fstree.WithDepth(0))
-		comp := &compression.Config{
-			Enabled: true,
-		}
-		fsTree.SetCompressor(comp)
 
 		// fstree must be initialized to create a descriptor
 		require.NoError(t, fsTree.Open(false))
