+++ pkg/local_object_storage/shard/put.go
@@ -91,11 +91,3 @@ func (s *Shard) Put(obj *object.Object, objBin []byte) error {
 
 	return nil
 }
-
-// NeedsCompression returns true if the object should be compressed.
-// For an object to be compressed 2 conditions must hold:
-// 1. Compression is enabled in settings.
-// 2. Object MIME Content-Type is allowed for compression.
-func (s *Shard) NeedsCompression(obj *object.Object) bool {
-	return s.compression.NeedsCompression(obj)
-}
